requirements.txt
streamlit
langchain
faiss-cpu
PyPDF2
langchain-community
langchain-groq
sentence-transformers
python-dotenv
